Move common logic to FileLocksmithLib
