+++ src/modules/FileLocksmith/FileLocksmithLib/AppLaunch.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <ShlObj.h>
+#include <string>
+#include <common/utils/process_path.h>
+
+#include "Constants.h"
+
+HRESULT LaunchUI(HMODULE module);
+
