+++ src/modules/FileLocksmith/FileLocksmithLib/framework.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
+
+#include <Windows.h>
+#include <Shlwapi.h>
+#include <ShlObj_core.h>
+#include <atlbase.h>
+#include <commctrl.h>
+
+// C++ Standard library
+#include <atomic>
+#include <string>
+#include <filesystem>
+#include <fstream>
